techpack: display: samsung: Limit max brightness level to 255
OneUI does not allow users to go above 255 unless adaptive brightness is enabled. Anything above 255 is considered as HBM, and even with Extra Brightness Mode enabled it goes up to 348 (588 nits) without adaptive brightness. Keeping the screen in HBM state all the time is dangerous for the hardware.

Exceeding the safe limits also causes issues with fingerprint HBM, as anything higher than 319 (525 nits) will cause the screen to go blank for a millisecond when leaving HBM, even in OneUI. The effects of this are more noticeable in AOSP (See 582112a62f2c).

Keep in mind that these values ​​are calculated with the Galaxy S20 FE (r8q) and may differ on other devices but in the end, result should be the same.

Change-Id: Ie389dadbfb6fe88e1a251c03c627173007fe4dcf